BACKEND/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.14",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.3.3",
    "pyjwt>=2.12.1",
    "pypdf>=6.9.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "reportlab>=4.4.10",
    "uvicorn>=0.42.0",
]
